# Experiment: Add an unsafe "borrow" unmarshaler that avoids some allocations