--- a/test/casttype/combos/unsafemarshaler/casttypepb_test.go
+++ b/test/casttype/combos/unsafemarshaler/casttypepb_test.go
@@ -124,8 +124,9 @@ func BenchmarkCastawayProtoUnmarshal(b *testing.B) {
 	msg := &Castaway{}
 	b.ResetTimer()
 	for i := 0; i < b.N; i++ {
-		total += len(datas[i%10000])
-		if err := github_com_gogo_protobuf_proto.Unmarshal(datas[i%10000], msg); err != nil {
+		data := datas[i%10000]
+		total += len(data)
+		if err := github_com_gogo_protobuf_proto.Unmarshal(data, msg); err != nil {
 			panic(err)
 		}
 	}
@@ -226,8 +227,9 @@ func BenchmarkWilsonProtoUnmarshal(b *testing.B) {
 	msg := &Wilson{}
 	b.ResetTimer()
 	for i := 0; i < b.N; i++ {
-		total += len(datas[i%10000])
-		if err := github_com_gogo_protobuf_proto.Unmarshal(datas[i%10000], msg); err != nil {
+		data := datas[i%10000]
+		total += len(data)
+		if err := github_com_gogo_protobuf_proto.Unmarshal(data, msg); err != nil {
 			panic(err)
 		}
 	}