This PR adds a new unsafe_borrow_unmarshaler tag that is more efficient on string or byte heavy objects during unmarshal, with the caveat that the caller must copy the buffer prior to passing it to Unmarshal() and that total memory allocated may go up.

The borrow unmarshaler avoids the []byte -> string cast Go performs to ensure strings are immutable via the unsafe package, relying on the caller to have provided an immutable buffer. For objects with 30-50% of fields being strings, this offers a substantial CPU improvement above 20% and a corresponding reduction in allocations.

The downside is that simple objects will not perform as well, and due to the signature of .Unmarshal() it is difficult to hide the "copy" operation on a per type basis (although not impossible, and could be done by splitting `UnmarshalBorrow()` and `Unmarshal()` for all types).

Added a new string heavy type and showed the improvement.  Many of the existing number heavy types were unchanged - so this is not a clear win for all types by far.

```
$ benchcmp /tmp/old /tmp/new
benchmark                                old ns/op     new ns/op     delta
BenchmarkStringHeavyProtoUnmarshal-8     945           728           -22.96%

benchmark                                old MB/s     new MB/s     speedup
BenchmarkStringHeavyProtoUnmarshal-8     412.41       535.12       1.30x

benchmark                                old allocs     new allocs     delta
BenchmarkStringHeavyProtoUnmarshal-8     14             8              -42.86%

benchmark                                old bytes     new bytes     delta
BenchmarkStringHeavyProtoUnmarshal-8     648           677           +4.48%
```

If this seems like a reasonable addition I can round out the remaining types to handle and add any necessary testing and doc.