@smarterclayton 2016-10-19
Implements / experiments with #174, but in a more useful direction (arena is much more complicated for arguably less benefit).