@awalterschulze 2016-10-19
First off I like this, but lets just discuss this a bit again.  Maybe there is a bigger speedup to be gained.
It seems what you are trying to achieve here is striving towards a zero copy unmarshaler.

Maybe you can give a use case where the user will always have copied the buffer before doing the Unmarshal?

You could probably also not copy any fixed64, fixed32, double or floats using unsafe?

How would you propose this split of Unmarshal and UnmarshalBorrow?