@smarterclayton 2016-10-20
> Maybe you can give a use case where the user will always have copied the buffer before doing the Unmarshal?

The user has code that wraps the use of unmarshal, usually.  I think the caller, to leverage zero-copy, has to know that he is using zero copy.  Ultimately any unmarshal tree that has any "zero-copy" unmarshaler has to be treated as completed zero-copy.

It's possible to avoid copying a fair number of values by direct byte manipulation.  However, I'm not sure that's actually faster in all cases since there is no allocation present.

For a split, I could imagine this marshaler providing the following signatures:

```
function (t *TYPE) Unmarshal(data []byte) error {
  copied := make([]byte, len(data))
  copy(copied, data)
  return t.UnmarshalBorrow(copied)
}

func (t *TYPE) UnmarshalBorrow(data []byte) error {
  ... normal generated borrower
}
```

Naive clients still call unmarshal and take the hit (but you can selective enable the borrow unmarshaler only for types that matter).  Sophisticated clients either call top level UnmarshalBorrow (because they have an immutable data object already) or just Unmarshal.  The generator would need to try to use the borrow unmarshaler if present all the way down.  However, if you had:

```
type 1 - supports borrow
  type 2 - doesn't support borrow
    type 3 - supports borrow
```

You'd end up doing two more copies than you need and end up losing the benefit.  So a warning / easy way to know that you're mixing borrow and unborrowed generators.