@awalterschulze 2016-10-22
Ok doesn't seem like my reply is getting through.
So here is a possible duplicate reply.

```
The user has code that wraps the use of unmarshal, usually. I think the caller, to leverage zero-copy, has to know that he is using zero copy. Ultimately any unmarshal tree that has any "zero-copy" unmarshaler has to be treated as completely zero-copy.
```

I am still trying to imagine some example where someone has a buffer that they are not reusing, because only then they will be able to use the borrow unmarshaler.  I am sure such a scenario exists.  I just want to know of one. 

```
It's possible to avoid copying a fair number of values by direct byte manipulation. However, I'm not sure that's actually faster in all cases since there is no allocation present.
```

Fair enough. Good point.

```
For a split, I could imagine this marshaler providing the following signatures:
```

I think this will slow down the original Unmarshal method, don't you? 

```
You'd end up doing two more copies than you need and end up losing the benefit. So a warning / easy way to know that you're mixing borrow and unborrowed generators.
```

Ok so lets rather have the UnmarshalBorrow method assume that its children also have an UnmarshalBorrow method.  Or are we going to type switch?  Not copying is probably a bigger time safer than a type assert is a time waster.