@smarterclayton 2016-10-22
GitHub is eating email responses over the last week or so - happening to me too.

> I am still trying to imagine some example where someone has a buffer that they are not reusing, because only then they will be able to use the borrow unmarshaler.

It would almost always be explicit (someone says copy) or naive (just using bad code).  So zero-copy tends to be something a user is conscious of.  However, the caller of `.Unmarshal*()` can sometimes know that the buffer is safe to use without copy (because due to higher level design in the program the buffer is unique, such as from `ioutil.ReadAll()` or from a frame decoding stream).  So there are cases where a copy is unnecessary because the use case always ends up with you having a uniquely referenced buffer.

> Or are we going to type switch? Not copying is probably a bigger time safer than a type assert is a time waster.

At the highest level, this could be looked at like a compiler optimization - given what we know about a type, is the borrow unmarshal worth it?  Some heuristics:

* Does the message contain enough strings/bytes relative to other fields to justify the cost of a copy (cost of copy is 1 allocation + CPU proportional to message size)
* Is the message included in another message that meets the above criteria?

The other tradeoff is code size - we could be time efficient by generating both paths (one that generates Unmarshal, and one that generates UnmarshalBorrow) but that would slow compilation and increase code size.  

I think in general that the message type alone determines whether borrow is worth it just based on the initial expectation.  I.e.:

```
// never worth generating borrow
message Foo {
  optional int32 field1 = 1
  optional int32 field2 = 2
}

// never worth generating borrow
message Bar {
  optional string field1 = 1
}

// probably always worth generating borrow
message Baz {
  repeated string field1 = 1
}

// probably always worth generating borrow
message Biz {
  repeated Bar field1 = 1
}
```