@awalterschulze 2016-10-23
> It would almost always be explicit (someone says copy) or naive (just using bad code). So zero-copy tends to be something a user is conscious of. However, the caller of .Unmarshal*() can sometimes know that the buffer is safe to use without copy (because due to higher level design in the program the buffer is unique, such as from ioutil.ReadAll() or from a frame decoding stream). So there are cases where a copy is unnecessary because the use case always ends up with you having a uniquely referenced buffer.

Ok cool so when using ioutil.ReadAll like in an http request.
Also maybe another is when you are reusing the input buffer, but once you have unmarshaled you process the message fully before unmarshaling the next message.

> At the highest level, this could be looked at like a compiler optimization - given what we know about a type, is the borrow unmarshal worth it? Some heuristics:

I think these heuristics should be left up to the user, since the user already has to make an education decision about when to use UnmarshalBorrow.  I don't think we can decide when to generate this for him. 

The fact that its up to the user's logic to decide when to use UnmarshalBorrow means than Unmarshal can never call into UnmarshalBorrow.

UnmarshalBorrow is never slower than Unmarshal, but in the case where the struct has no strings or bytes, the code is almost identical and increases the code size for seemingly no good reason.

But since Unmarshal can never call into UnmarshalBorrow, we need UnmarshalBorrow calls all the way down once the user has started calling UnmarshalBorrow.

An alternative is to have an `UnmarshalCopy(buf []byte, copy bool) error` method which can then be reused by both Unmarshal and UnmarshalBorrow.  This way we can UnmarshalBorrow all the way down and only when appropriate and code size is kept to a minimum when both Unmarshal and UnmarshalBorrow is generated.  Unfortunately when only one is generated there will be a if clause (for strings and bytes) which will never be executed, but this is probably an ok trade off.

customtype is a problem though, since it only has an `Unmarshal(buf []byte) error` method, but I guess we leave this and don't optimize this as well.